[Fix] #267 PR 리뷰 반영 수정 스토리지 파일 별도 분리
